+++ packages/antd-management-fast-framework/src/framework/DataSinglePageView/SinglePageModal/index.jsx
@@ -636,10 +636,12 @@ class SinglePageModal extends SinglePage {
         )}
         width={width}
         onCancel={this.onCancel}
-        bodyStyle={{
-          padding: 0,
-          overflowY: 'hidden',
-          overflowX: 'auto',
+        styles={{
+          body: {
+            padding: 0,
+            overflowY: 'hidden',
+            overflowX: 'auto',
+          },
         }}
         afterOpenChange={(v) => {
           that.doOtherWhenChangeVisible(v);